feat: adjustments to force parse feature
- Refactored git commits and release feature aswell to reduce duplicated code, as well as making it compatible with force parse

- Force parse is now compatible with the other parse options (initial and modified)

- Requires testing

+++ src/force.sh
@@ -9,24 +9,28 @@ features_plot=$5
 subfeatures_plot=$6
 metrics_plot=$7
 evolution_plot=$8
+modified_files=$9
 
 date="$(date +%Y-%m-%d_%H-%M-%S)"
 for file in $force_parse
 do
-    dir=$(dirname "$file")
-    outputFile=$(basename "$file" .owl)
-
-    # Remove folder if it exist. It will override results obtained on previous steps if the situation occurs.
-
-    rm -f $contents_folder/temp_results/$dir/$outputFile
-    mkdir -p $contents_folder/temp_results/$dir/$outputFile/$date/metrics
-    mkdir -p $contents_folder/temp_results/$dir/$outputFile/$date/img
-    outputFilePath="$contents_folder/temp_results/$dir/$outputFile/$date/metrics/$outputFile.xml"
-    java -jar $GITHUB_ACTION_PATH/libs/oquare-versions.jar --ontology "$file" --reasoner "$reasoner" --outputFile "$outputFilePath"
-
-    if [ -f "$outputFilePath" ]
+    if [ -z "$(printf '%s\n' "$modified_files" | grep -Fx "$file")" ]
     then
-        python $GITHUB_ACTION_PATH/src/main.py -i $contents_folder -s $dir -f $outputFile -d $date \
-            -M $model_plot -c $features_plot -S $subfeatures_plot -m $metrics_plot -e $evolution_plot
+        dir=$(dirname "$file")
+        outputFile=$(basename "$file" .owl)
+
+        # Remove folder if it exist. It will override results obtained on previous steps if the situation occurs.
+
+        rm -f $contents_folder/temp_results/$dir/$outputFile
+        mkdir -p $contents_folder/temp_results/$dir/$outputFile/$date/metrics
+        mkdir -p $contents_folder/temp_results/$dir/$outputFile/$date/img
+        outputFilePath="$contents_folder/temp_results/$dir/$outputFile/$date/metrics/$outputFile.xml"
+        java -jar $GITHUB_ACTION_PATH/libs/oquare-versions.jar --ontology "$file" --reasoner "$reasoner" --outputFile "$outputFilePath"
+
+        if [ -f "$outputFilePath" ]
+        then
+            python $GITHUB_ACTION_PATH/src/main.py -i $contents_folder -s $dir -f $outputFile -d $date \
+                -M $model_plot -c $features_plot -S $subfeatures_plot -m $metrics_plot -e $evolution_plot
+        fi
     fi
 done
